users_service/requirements.txt
fastapi
uvicorn

SQLAlchemy
psycopg2-binary
python-dotenv

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

PyJWT==2.6.0
python-multipart
pydantic[email]

memory_profiler